refactor(lib): rehome the digest and durable-write primitives into atomic-fs
Wish skills-everywhere-b deletes agent-sync.ts (Group 5) and
codex-activation-persistence.ts (Group 3). Both lend primitives to modules that
survive the wish, so the primitives move first, verbatim, into the leaf module
that already owns the crash-safety contract.

From agent-sync.ts: the whole physical-tree digest cluster --
computeDirDigest/computeFileDigest plus the private traversal helpers, and the
MANIFEST_NAME / PHYSICAL_TREE_IDENTITY_VERSION constants the digest grammar is
defined in terms of. agent-sync.ts imports the four it still uses internally
(computeExactDirDigest, computeLegacyRegularTreeDigest, physicalEntryKind,
PhysicalTreeEntry) and re-exports the four public ones, so doctor, uninstall,
runtime-integrations and the retirement module keep their current imports. Its
two remaining private hashFile calls now say computeFileDigest.

From codex-activation-persistence.ts: readBoundedRegularFile, fsyncParentDir,
atomicWriteFileSync and unlinkWithParentFsync with their private helpers. The
old module keeps renameNonOverwriting (Codex-only) and re-exports the four, so
install-version-marker.ts and update-capabilities.ts -- both survivors -- need
no change this wave. The BoundedFileRead / AtomicWriteOptions types are not
re-exported: no module outside that file ever imported them, and a re-export
with no importer is a knip failure. Both stay exported from atomic-fs.

Tests: the computeDirDigest describe block moves to atomic-fs.test.ts and the
four durable-write primitives gain coverage there (they had none). The
GENIE_HOME permission scan's exemption is re-keyed from a drifting
<path>:<line> to a content anchor, with a new test asserting every exemption
still matches exactly one real call site -- a stale line key failed OPEN, this
one fails closed; and its atomicWriteFileSync pin follows the call to
atomic-fs.ts.
