+++ src/lib/atomic-fs.ts
@@ -8,7 +8,7 @@
  */
 
 import { dlopen } from 'bun:ffi';
-import { randomBytes } from 'node:crypto';
+import { createHash, randomBytes } from 'node:crypto';
 import {
   constants,
   type Stats,
@@ -21,12 +21,22 @@ import {
   mkdirSync,
   openSync,
   readFileSync,
+  readSync,
   readdirSync,
+  readlinkSync,
   renameSync,
   rmSync,
   statSync,
+  unlinkSync,
   writeSync,
 } from 'node:fs';
+import { dirname, join, relative, sep } from 'node:path';
+
+/**
+ * `constants` under the name the persistence primitives moved in from
+ * `codex-activation-persistence.ts` already used, so those blocks read verbatim.
+ */
+const fsConstants = constants;
 
 /** Feature-detected glibc/musl sonames for the Linux renameat2 no-clobber fast path (x86_64). */
 const LINUX_LIBC_CANDIDATES = ['libc.so.6', 'ld-musl-x86_64.so.1', 'libc.musl-x86_64.so.1'] as const;
@@ -332,6 +342,325 @@ export function atomicRenameDirectoryNoClobber(stagedDir: string, targetDir: str
 
 function selectedNoClobberPlatform(deps: NoClobberDeps): NodeJS.Platform {
   return deps.platform ?? (deps.darwinOpener === undefined ? process.platform : 'darwin');
+}
+
+// ============================================================================
+// Physical-tree digest — moved verbatim out of `agent-sync.ts`
+// ============================================================================
+//
+// `computeDirDigest` / `computeFileDigest` are the verification primitive
+// `atomic-fs.test.ts` already used, and the only pieces of this cluster that
+// survive wish `skills-everywhere-b`. The private helpers travel with them
+// (a digest is defined by its traversal), and `MANIFEST_NAME` /
+// `PHYSICAL_TREE_IDENTITY_VERSION` come along because the digest grammar is
+// defined in terms of both. `agent-sync.ts` re-exports the public four and
+// imports back the four this move made non-private (`computeExactDirDigest`,
+// `computeLegacyRegularTreeDigest`, `physicalEntryKind`, `PhysicalTreeEntry`),
+// so no other consumer changes this wave.
+
+/** Manifest marker written into every managed skill dir. Exported: single source of truth. */
+export const MANIFEST_NAME = '.genie-sync.json';
+/** Physical-tree digest schema. Version 1 was the legacy regular-file content digest. */
+export const PHYSICAL_TREE_IDENTITY_VERSION = 2;
+// ============================================================================
+// Digest — a stable fingerprint of a physical directory tree
+// ============================================================================
+
+/**
+ * Version-2 SHA-256 identity over the root and every physical entry. Each entry
+ * contributes its normalized relative path, exact lstat kind, permission mode,
+ * and kind-specific payload (regular-file content hash or raw symlink target).
+ * Symlinks are never followed; FIFOs, sockets, devices, and other non-regular
+ * entries are represented rather than silently skipped. The manifest is always
+ * excluded because its digest field would otherwise be self-referential.
+ */
+export function computeDirDigest(dir: string, exclude?: Set<string>): string {
+  const excluded = new Set([...(exclude ?? [])].map(normalizePhysicalRelPath));
+  excluded.add(MANIFEST_NAME);
+  return computePhysicalTreeDigest(dir, excluded);
+}
+
+export function computeExactDirDigest(dir: string): string {
+  return computePhysicalTreeDigest(dir, new Set());
+}
+
+function computePhysicalTreeDigest(dir: string, excluded: Set<string>): string {
+  const rootStat = lstatSync(dir);
+  if (!rootStat.isDirectory() || rootStat.isSymbolicLink()) {
+    throw new Error(`physical tree root is not a directory: ${dir}`);
+  }
+  const entries: PhysicalTreeEntry[] = [physicalTreeEntry('.', rootStat, dir)];
+  collectPhysicalTreeEntries(dir, dir, excluded, entries);
+  entries.sort(byRel);
+  const digest = createHash('sha256');
+  digest.update(`genie-physical-tree-v${PHYSICAL_TREE_IDENTITY_VERSION}\0`);
+  for (const entry of entries) {
+    updateLengthPrefixed(digest, entry.rel);
+    updateLengthPrefixed(digest, entry.kind);
+    updateLengthPrefixed(digest, entry.mode.toString(8));
+    updateLengthPrefixed(digest, entry.payload);
+  }
+  return digest.digest('hex');
+}
+
+export interface PhysicalTreeEntry {
+  rel: string;
+  kind: 'directory' | 'file' | 'symlink' | 'fifo' | 'socket' | 'block-device' | 'character-device' | 'other';
+  mode: number;
+  payload: string;
+}
+
+function byRel(a: { rel: string }, b: { rel: string }): number {
+  if (a.rel < b.rel) return -1;
+  if (a.rel > b.rel) return 1;
+  return 0;
+}
+
+function normalizePhysicalRelPath(path: string): string {
+  return path.split(sep).join('/');
+}
+
+export function physicalEntryKind(stat: Stats): PhysicalTreeEntry['kind'] {
+  if (stat.isSymbolicLink()) return 'symlink';
+  if (stat.isDirectory()) return 'directory';
+  if (stat.isFile()) return 'file';
+  if (stat.isFIFO()) return 'fifo';
+  if (stat.isSocket()) return 'socket';
+  if (stat.isBlockDevice()) return 'block-device';
+  if (stat.isCharacterDevice()) return 'character-device';
+  return 'other';
+}
+
+function physicalTreeEntry(rel: string, stat: Stats, absolute: string): PhysicalTreeEntry {
+  const kind = physicalEntryKind(stat);
+  const payload = kind === 'file' ? hashFile(absolute) : kind === 'symlink' ? readlinkSync(absolute) : '';
+  return { rel, kind, mode: stat.mode & 0o7777, payload };
+}
+
+function collectPhysicalTreeEntries(
+  root: string,
+  current: string,
+  excluded: Set<string>,
+  out: PhysicalTreeEntry[],
+): void {
+  for (const name of readdirSync(current)) {
+    const abs = join(current, name);
+    const rel = normalizePhysicalRelPath(relative(root, abs));
+    if (excluded.has(rel)) continue;
+    const stat = lstatSync(abs);
+    const entry = physicalTreeEntry(rel, stat, abs);
+    out.push(entry);
+    if (entry.kind === 'directory') collectPhysicalTreeEntries(root, abs, excluded, out);
+  }
+}
+
+function updateLengthPrefixed(digest: ReturnType<typeof createHash>, value: string): void {
+  const bytes = Buffer.from(value);
+  digest.update(String(bytes.length));
+  digest.update(':');
+  digest.update(bytes);
+  digest.update('\0');
+}
+
+/**
+ * Legacy v1 digest, accepted only when every physical entry is a regular file
+ * or directory. A symlink or special entry in an old tree therefore revokes
+ * deletion/update authority instead of recreating the legacy follow/skip bug.
+ */
+export function computeLegacyRegularTreeDigest(dir: string): string | null {
+  const rootStat = lstatSync(dir);
+  if (!rootStat.isDirectory() || rootStat.isSymbolicLink()) return null;
+  const files: Array<{ rel: string; hash: string }> = [];
+  const visit = (current: string): boolean => {
+    for (const name of readdirSync(current)) {
+      const absolute = join(current, name);
+      const rel = relative(dir, absolute);
+      if (rel === MANIFEST_NAME) continue;
+      const stat = lstatSync(absolute);
+      if (stat.isDirectory() && !stat.isSymbolicLink()) {
+        if (!visit(absolute)) return false;
+      } else if (stat.isFile() && !stat.isSymbolicLink()) {
+        files.push({ rel, hash: hashFile(absolute) });
+      } else {
+        return false;
+      }
+    }
+    return true;
+  };
+  if (!visit(dir)) return null;
+  files.sort(byRel);
+  const digest = createHash('sha256');
+  for (const file of files) {
+    digest.update(file.rel);
+    digest.update('\0');
+    digest.update(file.hash);
+    digest.update('\0');
+  }
+  return digest.digest('hex');
+}
+
+function hashFile(path: string): string {
+  return createHash('sha256').update(readFileSync(path)).digest('hex');
+}
+
+export function computeFileDigest(path: string): string {
+  return hashFile(path);
+}
+
+// ============================================================================
+// Bounded reads + durable writes — moved verbatim out of
+// `codex-activation-persistence.ts`
+// ============================================================================
+//
+// Nothing in these four primitives is Codex-specific: they are bounded
+// regular-file reads that fail closed on symlink/oversize, a best-effort
+// parent-directory fsync, an atomic backup-first write, and an idempotent
+// unlink. Wish `skills-everywhere-b` Group 3 deletes their old home, and two
+// consumers that survive it -- `install-version-marker.ts` and
+// `update-capabilities.ts` -- depend on them, so they live here with the rest
+// of the crash-safety contract. The old module re-exports them for this wave.
+
+/** O_NOFOLLOW is POSIX-only; degrade to 0 on platforms that lack it. */
+const O_NOFOLLOW = (fsConstants.O_NOFOLLOW ?? 0) as number;
+
+export type BoundedFileRead =
+  | { status: 'ok'; content: string; size: number }
+  | { status: 'absent' }
+  | { status: 'symlink' }
+  | { status: 'non-regular' }
+  | { status: 'oversized'; size: number }
+  | { status: 'unreadable'; detail: string };
+
+/**
+ * Read a regular file bounded to `maxBytes`, following no symlink at the final
+ * component. A symlink, non-regular kind, or oversize is a distinct fail-closed
+ * category the caller must handle explicitly; nothing here is mutated.
+ */
+export function readBoundedRegularFile(path: string, maxBytes: number): BoundedFileRead {
+  let stat: Stats;
+  try {
+    stat = lstatSync(path);
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException).code === 'ENOENT') return { status: 'absent' };
+    return { status: 'unreadable', detail: errorText(error) };
+  }
+  if (stat.isSymbolicLink()) return { status: 'symlink' };
+  if (!stat.isFile()) return { status: 'non-regular' };
+  if (stat.size > maxBytes) return { status: 'oversized', size: stat.size };
+  try {
+    const fd = openSync(path, fsConstants.O_RDONLY | O_NOFOLLOW);
+    try {
+      const buffer = Buffer.alloc(stat.size);
+      let read = 0;
+      while (read < stat.size) {
+        const chunk = readSync(fd, buffer, read, stat.size - read, read);
+        if (chunk <= 0) break;
+        read += chunk;
+      }
+      return { status: 'ok', content: buffer.subarray(0, read).toString('utf8'), size: read };
+    } finally {
+      closeSync(fd);
+    }
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException).code === 'ELOOP') return { status: 'symlink' };
+    return { status: 'unreadable', detail: errorText(error) };
+  }
+}
+
+/** Best-effort parent-directory fsync; unsupported filesystems degrade silently. */
+export function fsyncParentDir(path: string): void {
+  try {
+    const dirFd = openSync(dirname(path), 'r');
+    try {
+      fsyncSync(dirFd);
+    } finally {
+      closeSync(dirFd);
+    }
+  } catch {
+    // Directory fsync is not portable; the file fsync + atomic rename remain sound.
+  }
+}
+
+export interface AtomicWriteOptions {
+  mode?: number;
+  /** Copy an existing regular target to a timestamped sidecar before replacing it. */
+  backup?: boolean;
+}
+
+/**
+ * Atomically publish `content` to `path`: create the parent, back up any prior
+ * regular file, write a private staging sibling, fsync it, rename over the
+ * target, and fsync the parent directory. The rename is the commit point, so a
+ * crash before it leaves the prior file intact.
+ */
+export function atomicWriteFileSync(path: string, content: string, options: AtomicWriteOptions = {}): void {
+  const mode = options.mode ?? 0o600;
+  const dir = dirname(path);
+  // Every caller writes under GENIE_HOME — the activation store's paths derive
+  // from `resolveGenieHome()`, delivery evidence sits under
+  // `<GENIE_HOME>/<evidence>`, and the capability sidecar lands beside the
+  // prior binary in `<GENIE_HOME>/bin/.previous`. A direct mode is therefore
+  // correct here; no caller opt-in is needed.
+  mkdirSync(dir, { recursive: true, mode: 0o700 });
+  if (options.backup) backupExistingRegularFile(path);
+  const staging = join(dir, `.${basenameOf(path)}.staging-${process.pid}-${uniqueSuffix()}`);
+  const fd = openSync(staging, fsConstants.O_WRONLY | fsConstants.O_CREAT | fsConstants.O_EXCL, mode);
+  try {
+    const buffer = Buffer.from(content, 'utf8');
+    let written = 0;
+    while (written < buffer.length) {
+      written += writeSync(fd, buffer, written, buffer.length - written, null);
+    }
+    fsyncSync(fd);
+  } finally {
+    closeSync(fd);
+  }
+  renameSync(staging, path);
+  fsyncParentDir(path);
+}
+
+/** Delete a file and fsync its parent; ENOENT is treated as already-released. */
+export function unlinkWithParentFsync(path: string): void {
+  try {
+    unlinkSync(path);
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException).code !== 'ENOENT') throw error;
+  }
+  fsyncParentDir(path);
+}
+
+function backupExistingRegularFile(path: string): void {
+  let stat: Stats;
+  try {
+    stat = lstatSync(path);
+  } catch {
+    return;
+  }
+  if (!stat.isFile() || stat.isSymbolicLink()) return;
+  const backup = `${path}.genie-backup-${timestamp()}-${uniqueSuffix()}`;
+  try {
+    copyFileSync(path, backup);
+    fsyncParentDir(backup);
+  } catch {
+    // A best-effort backup failure must not block the durable write itself.
+  }
+}
+
+function basenameOf(path: string): string {
+  const parts = path.split(/[\\/]/);
+  return parts[parts.length - 1] || 'state';
+}
+
+function timestamp(): string {
+  return new Date().toISOString().replace(/[:.]/g, '-');
+}
+
+function uniqueSuffix(): string {
+  return randomBytes(6).toString('hex');
+}
+
+function errorText(error: unknown): string {
+  return error instanceof Error ? error.message : String(error);
 }
 
 export {
