+++ src/lib/codex-activation-persistence.ts
@@ -12,124 +12,24 @@
  * distinguish physical fault categories (absent / symlink / non-regular /
  * oversized / unreadable) without mutation, so callers can fail closed on an
  * exact category rather than collapsing every fault into "missing".
+ *
+ * The four generic primitives -- `readBoundedRegularFile`, `fsyncParentDir`,
+ * `atomicWriteFileSync` and `unlinkWithParentFsync` -- now live in
+ * `atomic-fs.ts`: two of their consumers (`install-version-marker.ts`,
+ * `update-capabilities.ts`) survive wish `skills-everywhere-b`, which deletes
+ * this module in Group 3. They are re-exported here so no caller changes this
+ * wave; `renameNonOverwriting` stays because only Codex activation uses it.
+ *
+ * The `BoundedFileRead` and `AtomicWriteOptions` types are NOT re-exported: no
+ * module outside this one ever imported them (they were reachable only because
+ * this file used them locally), and a pure type re-export with no importer is a
+ * knip failure. Both are still exported from `atomic-fs.ts`.
  */
 
-import { randomBytes } from 'node:crypto';
-import {
-  type Stats,
-  closeSync,
-  copyFileSync,
-  constants as fsConstants,
-  fsyncSync,
-  linkSync,
-  lstatSync,
-  mkdirSync,
-  openSync,
-  readSync,
-  renameSync,
-  unlinkSync,
-  writeSync,
-} from 'node:fs';
-import { dirname, join } from 'node:path';
+import { linkSync, unlinkSync } from 'node:fs';
+import { atomicWriteFileSync, fsyncParentDir, readBoundedRegularFile, unlinkWithParentFsync } from './atomic-fs.js';
 
-/** O_NOFOLLOW is POSIX-only; degrade to 0 on platforms that lack it. */
-const O_NOFOLLOW = (fsConstants.O_NOFOLLOW ?? 0) as number;
-
-export type BoundedFileRead =
-  | { status: 'ok'; content: string; size: number }
-  | { status: 'absent' }
-  | { status: 'symlink' }
-  | { status: 'non-regular' }
-  | { status: 'oversized'; size: number }
-  | { status: 'unreadable'; detail: string };
-
-/**
- * Read a regular file bounded to `maxBytes`, following no symlink at the final
- * component. A symlink, non-regular kind, or oversize is a distinct fail-closed
- * category the caller must handle explicitly; nothing here is mutated.
- */
-export function readBoundedRegularFile(path: string, maxBytes: number): BoundedFileRead {
-  let stat: Stats;
-  try {
-    stat = lstatSync(path);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code === 'ENOENT') return { status: 'absent' };
-    return { status: 'unreadable', detail: errorText(error) };
-  }
-  if (stat.isSymbolicLink()) return { status: 'symlink' };
-  if (!stat.isFile()) return { status: 'non-regular' };
-  if (stat.size > maxBytes) return { status: 'oversized', size: stat.size };
-  try {
-    const fd = openSync(path, fsConstants.O_RDONLY | O_NOFOLLOW);
-    try {
-      const buffer = Buffer.alloc(stat.size);
-      let read = 0;
-      while (read < stat.size) {
-        const chunk = readSync(fd, buffer, read, stat.size - read, read);
-        if (chunk <= 0) break;
-        read += chunk;
-      }
-      return { status: 'ok', content: buffer.subarray(0, read).toString('utf8'), size: read };
-    } finally {
-      closeSync(fd);
-    }
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code === 'ELOOP') return { status: 'symlink' };
-    return { status: 'unreadable', detail: errorText(error) };
-  }
-}
-
-/** Best-effort parent-directory fsync; unsupported filesystems degrade silently. */
-export function fsyncParentDir(path: string): void {
-  try {
-    const dirFd = openSync(dirname(path), 'r');
-    try {
-      fsyncSync(dirFd);
-    } finally {
-      closeSync(dirFd);
-    }
-  } catch {
-    // Directory fsync is not portable; the file fsync + atomic rename remain sound.
-  }
-}
-
-export interface AtomicWriteOptions {
-  mode?: number;
-  /** Copy an existing regular target to a timestamped sidecar before replacing it. */
-  backup?: boolean;
-}
-
-/**
- * Atomically publish `content` to `path`: create the parent, back up any prior
- * regular file, write a private staging sibling, fsync it, rename over the
- * target, and fsync the parent directory. The rename is the commit point, so a
- * crash before it leaves the prior file intact.
- */
-export function atomicWriteFileSync(path: string, content: string, options: AtomicWriteOptions = {}): void {
-  const mode = options.mode ?? 0o600;
-  const dir = dirname(path);
-  // Every caller writes under GENIE_HOME — the activation store's paths derive
-  // from `resolveGenieHome()`, delivery evidence sits under
-  // `<GENIE_HOME>/<evidence>`, and the capability sidecar lands beside the
-  // prior binary in `<GENIE_HOME>/bin/.previous`. A direct mode is therefore
-  // correct here; no caller opt-in is needed.
-  mkdirSync(dir, { recursive: true, mode: 0o700 });
-  if (options.backup) backupExistingRegularFile(path);
-  const staging = join(dir, `.${basenameOf(path)}.staging-${process.pid}-${uniqueSuffix()}`);
-  const fd = openSync(staging, fsConstants.O_WRONLY | fsConstants.O_CREAT | fsConstants.O_EXCL, mode);
-  try {
-    const buffer = Buffer.from(content, 'utf8');
-    let written = 0;
-    while (written < buffer.length) {
-      written += writeSync(fd, buffer, written, buffer.length - written, null);
-    }
-    fsyncSync(fd);
-  } finally {
-    closeSync(fd);
-  }
-  renameSync(staging, path);
-  fsyncParentDir(path);
-}
+export { atomicWriteFileSync, fsyncParentDir, readBoundedRegularFile, unlinkWithParentFsync };
 
 /**
  * Rename `from` to `to` without ever overwriting an existing `to`. Used for
@@ -158,16 +58,6 @@ export function renameNonOverwriting(from: string, to: string): { moved: boolean
   }
 }
 
-/** Delete a file and fsync its parent; ENOENT is treated as already-released. */
-export function unlinkWithParentFsync(path: string): void {
-  try {
-    unlinkSync(path);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== 'ENOENT') throw error;
-  }
-  fsyncParentDir(path);
-}
-
 function linkAndUnlink(from: string, to: string): void {
   // renameSync overwrites on POSIX, so use link (fails EEXIST) then unlink the
   // source. This preserves the "never overwrite" invariant atomically.
@@ -177,38 +67,4 @@ function linkAndUnlink(from: string, to: string): void {
   } catch {
     // The hard link is committed; a lingering source is harmless.
   }
-}
-
-function backupExistingRegularFile(path: string): void {
-  let stat: Stats;
-  try {
-    stat = lstatSync(path);
-  } catch {
-    return;
-  }
-  if (!stat.isFile() || stat.isSymbolicLink()) return;
-  const backup = `${path}.genie-backup-${timestamp()}-${uniqueSuffix()}`;
-  try {
-    copyFileSync(path, backup);
-    fsyncParentDir(backup);
-  } catch {
-    // A best-effort backup failure must not block the durable write itself.
-  }
-}
-
-function basenameOf(path: string): string {
-  const parts = path.split(/[\\/]/);
-  return parts[parts.length - 1] || 'state';
-}
-
-function timestamp(): string {
-  return new Date().toISOString().replace(/[:.]/g, '-');
-}
-
-function uniqueSuffix(): string {
-  return randomBytes(6).toString('hex');
-}
-
-function errorText(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
 }
